pyproject.toml
[project]
name = "knitty"
version = "0.1.0"
description = "An intelligent CV tailoring system that optimizes resumes for specific job postings"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "python-dotenv>=1.0.0",
    "faiss-cpu>=1.12.0",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.33",
    "langchain-core>=0.3.0",
    "nest-asyncio>=1.6.0",
    "playwright>=1.55.0",
    "pypdf>=6.1.0",
    "numpy>=1.24.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "streamlit>=1.32.0",
    "python-multipart>=0.0.6",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "pip>=25.2",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]
